move CfgEnv
