+++ crates/context-interface/Cargo.toml
@@ -1,0 +1,4 @@
+[features]
+# Minimal features needed for the interface
+default = ["std"]
+std = []
