+++ crates/context/Cargo.toml
@@ -33,6 +33,7 @@ bytecode.workspace = true
 
 # misc
 derive-where.workspace = true
+cfg-if.workspace = true
 
 # Optional
 serde = { version = "1.0", default-features = false, features = [
@@ -44,13 +45,20 @@ serde = { version = "1.0", default-features = false, features = [
 database.workspace = true
 
 [features]
+# Implementation-specific features
 default = ["std"]
-std = ["serde?/std"]
-serde = [
-    "dep:serde",
-    "primitives/serde",
-    "specification/serde",
-    "state/serde",
-    "context-interface/serde",
+std = []
+dev = [
+    "memory_limit",
+    "optional_balance_check",
+    "optional_block_gas_limit",
+    "optional_eip3607",
+    "optional_gas_refund",
+    "optional_no_base_fee",
 ]
-serde-json = ["serde"]
+memory_limit = []
+optional_balance_check = []
+optional_block_gas_limit = []
+optional_eip3607 = []
+optional_gas_refund = []
+optional_no_base_fee = []
