+++ crates/context/interface/Cargo.toml
@@ -29,7 +29,6 @@ state.workspace = true
 specification.workspace = true
 
 # mics
-cfg-if.workspace = true
 auto_impl.workspace = true
 
 # Optional
@@ -47,20 +46,3 @@ std = ["serde?/std"]
 serde = ["dep:serde", "primitives/serde", "specification/serde", "state/serde"]
 serde-json = ["serde"]
 
-# Enable additional features for development
-# They add functions to Cfg trait.
-dev = [
-    "memory_limit",
-    "optional_balance_check",
-    "optional_block_gas_limit",
-    "optional_eip3607",
-    "optional_gas_refund",
-    "optional_no_base_fee",
-
-]
-memory_limit = []
-optional_balance_check = []
-optional_block_gas_limit = []
-optional_eip3607 = []
-optional_gas_refund = []
-optional_no_base_fee = []
