+++ crates/revm/Cargo.toml
@@ -82,12 +82,12 @@ dev = [
     "optional_gas_refund",
     "optional_no_base_fee",
 ]
-memory_limit = ["context-interface/memory_limit", "interpreter/memory_limit"]
-optional_balance_check = ["context-interface/optional_balance_check"]
-optional_block_gas_limit = ["context-interface/optional_block_gas_limit"]
-optional_eip3607 = ["context-interface/optional_eip3607"]
-optional_gas_refund = ["context-interface/optional_gas_refund"]
-optional_no_base_fee = ["context-interface/optional_no_base_fee"]
+memory_limit = ["context/memory_limit", "interpreter/memory_limit"]
+optional_balance_check = ["context/optional_balance_check"]
+optional_block_gas_limit = ["context/optional_block_gas_limit"]
+optional_eip3607 = ["context/optional_eip3607"]
+optional_gas_refund = ["context/optional_gas_refund"]
+optional_no_base_fee = ["context/optional_no_base_fee"]
 
 # See comments in `precompile`
 secp256k1 = ["precompile/secp256k1"]
